SITAM CONFERENCE README.md
As an achievement in the Electrical and Electronics Engineering (EEE) department, I had the privilege of developing the official conference website for SITAM-2024. The primary goal of this website was to effectively showcase the event details, such as the conference agenda, speaker information, registration process, and institutional pride, while also enhancing engagement with participants and attendees. The website was designed with a user-friendly interface, ensuring that users could easily navigate through different sections and access relevant information with minimal effort.