SITAM CONFERENCE
As an achievement in the EEE department, I developed a conference website for SITAM-2024. The website effectively showcases event details and highlights institutional pride. Designed with a user-friendly interface, it ensures seamless navigation and accessibility. It enhances engagement by providing comprehensive information about the conference. This project reflects my technical skills and dedication to academic excellence.